pyproject.toml
[project]
name = "clean-architecture"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.15.2",
    "bcrypt>=4.3.0",
    "email-validator>=2.2.0",
    "fastapi>=0.115.12",
    "opentelemetry-distro>=0.54b1",
    "opentelemetry-exporter-otlp-proto-grpc>=1.33.1",
    "opentelemetry-instrumentation-asgi==0.54b1",
    "opentelemetry-instrumentation-asyncio==0.54b1",
    "opentelemetry-instrumentation-click==0.54b1",
    "opentelemetry-instrumentation-dbapi==0.54b1",
    "opentelemetry-instrumentation-fastapi==0.54b1",
    "opentelemetry-instrumentation-httpx==0.54b1",
    "opentelemetry-instrumentation-logging==0.54b1",
    "opentelemetry-instrumentation-psycopg2==0.54b1",
    "opentelemetry-instrumentation-sqlalchemy==0.54b1",
    "opentelemetry-instrumentation-sqlite3==0.54b1",
    "opentelemetry-instrumentation-starlette==0.54b1",
    "opentelemetry-instrumentation-threading==0.54b1",
    "opentelemetry-instrumentation-tortoiseorm==0.54b1",
    "opentelemetry-instrumentation-urllib==0.54b1",
    "opentelemetry-instrumentation-wsgi==0.54b1",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.2",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "httpx>=0.28.1",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "ruff>=0.11.10",
]
